pyproject.toml
[project]
name = "otoposuto"
version = "0.1.0"
description = ""
authors = [
    {name = "アリメクタ アレネクト",email = "alimektor@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)"
]

[tool.poetry.scripts]
otoposuto = "otoposuto.main:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
